+++ app/auth/service/user.py
@@ -1,8 +1,8 @@
 #!/usr/bin/python3
 # -*- coding: utf-8 -*-
 
-from app.auth.models import VSUser, VSRole, VSDepartment
-from lib.database import DataTable
+from app.auth.models import VSUser, VSRole
+from app.database import DataTable
 import json
 
 
@@ -22,24 +22,14 @@ class UserService:
             data = user.to_dict()
 
             # 查询部门所属信息
-            if user.department_id:
-                department = VSDepartment.query.filter_by(id=user.department_id).first_or_404()
-                if department:
-                    data['department'] = department.to_dict()
-                else:
-                    data['department'] = {}
-            else:
-                data['department'] = {}
+            data['department'] = {}
+            if user.department:
+                data['department'] = user.department.to_dict()
 
             # 查询所属角色信息
-            if user.role_id:
-                role = VSRole.query.filter_by(id=user.role_id).first_or_404()
-                if role:
-                    data['role'] = role.to_dict()
-                else:
-                    data['role'] = {}
-            else:
-                data['role'] = {}
+            data['role'] = {}
+            if user.role:
+                data['role'] = user.role.to_dict()
         else:
             return 50000, None, '账户基本信息不存在'
 
@@ -52,17 +42,33 @@ class UserService:
         :param request: http request
         :return:
         """
-        datatable = DataTable(
+        table = DataTable(
             model=VSUser,
             columns=[VSUser.id, VSUser.username, VSUser.email],
-            sortable=[],
+            sortable=[VSUser.id],
             searchable=[VSUser.username],
-            filterable=[VSUser.active],
+            filterable=[VSUser.company_id, VSUser.department_id, VSUser.role_id],
             limits=[25, 50, 100],
             request=request
         )
 
-        return 20000, datatable.query.items(), '账户信息列表查询成功'
+        data = []
+        users = table.items()
+        for user in users:
+            user_dict = user.to_dict()
+            # 查询部门所属信息
+            user_dict['department'] = {}
+            if user.department:
+                user_dict['department'] = user.department.to_dict()
+
+            # 查询所属角色信息
+            user_dict['role'] = {}
+            if user.role:
+                user_dict['role'] = user.role.to_dict()
+
+            data.append(user_dict)
+
+        return 20000, data, '账户信息列表查询成功'
 
     @classmethod
     def update(cls, id, request) -> (bool, dict, str):
@@ -77,7 +83,21 @@ class UserService:
             args_data = json.loads(request.data)
 
         user = VSUser.query.filter_by(id=id).first_or_404()
-        return 20000, user.update(**args_data), '账户信息更新成功'
+        if user:
+            new_user = user.update(**args_data)
+            data = new_user.to_dict()
+            # 查询部门所属信息
+            data['department'] = {}
+            if user.department:
+                data['department'] = user.department.to_dict()
+
+            # 查询所属角色信息
+            data['role'] = {}
+            if user.role:
+                data['role'] = user.role.to_dict()
+        else:
+            return 50000, {}, '账户基本信息更新失败'
+        return 20000, data, '账户信息更新成功'
 
     @classmethod
     def create(cls, request) -> (bool, dict, str):
@@ -91,7 +111,23 @@ class UserService:
         else:
             args_data = json.loads(request.data)
 
-        return 20000, VSUser.create(**args_data), '账户信息保存成功'
+        user = VSUser.create(**args_data)
+        if user:
+            data = user.to_dict()
+
+            # 查询部门所属信息
+            data['department'] = {}
+            if user.department:
+                data['department'] = user.department.to_dict()
+
+            # 查询所属角色信息
+            data['role'] = {}
+            if user.role:
+                data['role'] = user.role.to_dict()
+
+            return 20000, data, '账户信息保存成功'
+
+        return 20000, None, '账户信息保存失败'
 
     @classmethod
     def delete(cls, id: int) -> (bool, dict, str):
@@ -99,7 +135,10 @@ class UserService:
         :param id 账户id
         """
         user = VSUser.query.filter_by(id=id).first_or_404()
-        return 20000, user.delete(), '账户信息删除成功'
+        if user:
+            return 20000, user.delete(), '账户信息删除成功'
+
+        return 50000, None, '账户基本信息删除失败'
 
     @classmethod
     def login(cls, request) -> (bool, dict, str):
@@ -141,13 +180,12 @@ class UserService:
         return result, data, message
 
     @classmethod
-    def logout(cls, username: str) -> (bool, dict, str):
+    def logout(cls) -> (bool, dict, str):
         """
         退出登录
-        :param username: 账户名称
         :return:
         """
-        return 20000, username, '退出登录成功'
+        return 20000, None, '退出登录成功'
 
     @classmethod
     def register(cls, request) -> (bool, dict, str):
@@ -157,12 +195,3 @@ class UserService:
             args_data = json.loads(request.data)
 
         return 20000, VSUser.create(**args_data), '注册成功'
-
-    @classmethod
-    def update_role(cls, request) -> (bool, dict, str):
-        if request.is_json:
-            args_data = request.get_json()
-        else:
-            args_data = json.loads(request.data)
-
-        return 20000, VSUser.create(**args_data), '更新角色成功'
